Move the interprocess messaging class to a separate file.
Visual Studio insists that if a class inherits from a Winforms Form,
it must be the first in the source file, because it lets it add metadata
for the GUI editor, which we don’t even use because the form is never
displayed (its only purpose is to receive messages).

Also, the class will grow larger and the sooner I split it the better.