Finish implementing keyboard LED support.
If the key used as the compose key normally has a LED indicator, such as
the Caps Lock or Num Lock keys, then we turn on the indicator during the
compose sequence. Fixes #82.

Caveats: we check for the key status each time the compose key is
pressed or released, which may interfer with other processes that change
the LED indicator status. Probably not a big deal.