Fix issues with the “preserve compose key” feature.
When the Shift key was pressed while Compose was down, we would disable
Compose in case Compose+Shift was part of a combination, for instance
Ctrl-Shift-n to create a new window in some applications. Unfortunately
this could also mean Compose + capital N. The latter is a lot more
likely, so we must absolutely not abort composition here.

Solution: we disable the feature if the current key is a modifier key,
and we only consider key down events. Closes #101 #105.