Work around a bug in mintty.exe.
Explanation of the bug: when RAlt serves as the compose key, one can
press RAlt + Tab in order to cycle through windows. WinCompose emulates
this with keybd_event() by manually sending KEYDOWN and KEYUP for RAlt.
However the current application does not seem to get the KEYUP event
and remains stuck in a state where the Alt key is pressed.

Sending KEYUP for Left Alt when Right Alt is the compose key seems to
fix the problem without any apparent side effect.