Fix behaviour when AltGr is the Compose key. Fixes #39 #80 #114.
If AltGr is pressed and released, it counts as a regular Compose key press.

If AltGr is down and a key is pressed, and the key combination is a standard
AltGr combination, we switch to regular AltGr behaviour and it does not count
as a Compose key press.

If AltGr is down and a key is pressed but it has no effect with AltGr, then
it adds both <Multi_key> and the second key to the compose sequence.