When typing certain Unicode sequences, stop after 4 digits
When typing U+0xxx, it can be assumed that a four-digit code point is
meant. This is consistent with the notational conventions in the Unicode
standard[1] (which technically speaking only apply to that document),
saying that code points above U+FFFF are not written with leading zeros.

Since the planes 3 to 13 are unassigned, four-digit code points starting
with one of these numbers cannot currently lead to a valid code point,
and probably will not do so in the foreseeable future, since there is
still plenty of space in the planes 1 and 2. Stopping after 4 digits
makes typing CJK code points easier.

Fixes #119.

[1] http://www.unicode.org/versions/Unicode8.0.0/appA.pdf