Add support for Inno Setup 5.5.7.
From the changelog: “As recommended by Microsoft's desktop applications
guideline, DisableWelcomePage now defaults to yes. Additionally DisableDirPage
and DisableProgramGroupPage now default to auto. The defaults in all previous
versions were no.”

This means that we have to insert a new hidden page instead of relying on
the wpDirPage, because Inno Setup may decide to skip it entirely.