Start working on a bug where some keys seem stuck.
There was at least one code path where KeyUp event were not
properly sent, but this is not enough, we need to make the
AltGr detection code more robust, or we will forget to send
KeyUp events when disabling combination mode. This commit also
adds some improved information to the debug messages.