Use Git submodules for third-party data.
Instead of downloading data using wget or relying on the Xorg compose
file being available on a Linux computer, we add these projects as Git
submodules (they all use Git) and simply refer to the source files.

Additionally, we embed those files as resources instead of copying
them around. The downside is that the user cannot hack them locally.
The upside is that the user cannot hack them locally.