Changed default install location
The installer is broken when installing to the default location for non-admin users. The default location (Program Files) requires admin access, so the installer prompts for elevation. If admin credentials are entered, the installer runs successfully, but only the admin account gets app shortcuts and registry entries, not the standard user account that initiated the install. Changing the default location to the local AppData folder will ensure that non-admin users can install without admin credentials, with shortcuts, run entries, and uninstallers fully working.