Automatically clean up the notification area.
This is a workaround for a decades-long annoying Windows behaviour, where
zombie icons would remain in the notification area when programs stopped
unexpectedly, or without hiding their icon beforehands.

This is really not a job WinCompose should be doing, but I don’t feel
like installing extra software for such a meager task.