Work around interference from the IME system in ToUnicode(). See #210.
For instance when the Japanese IME is activated, it could happen that
”Compose a e” would result in “ﾁe” instead of “æ” because ToUnicode()
would yield “ﾁ” instead of “a”.

Our workaround is to pretend we use an US QWERTY keyboard when the
Japanese IME input is detected. There may be other input systems that
exhibit the same symptoms, so they’ll have to be revisited, too.