Fix the icon list in the final application executable; the first one was missing.
This addresses the missing icon issue reported in #306 but we still have the
unresolved issue of the different windows all showing the same icon in the
task bar. I believe this is due to these icons being part of the same process,
so the explorer caches their icons under the same key.