Update window content when reloading user-defined sequences
The sequence list is loaded into the RootViewModel on construction, so
the sequences in the Settings don't stay in sync after that. When
reloading the user-defined sequences we therefore need to create a new
RootViewModel for the UI to reflect the active list of sequences.

Steps to reproduce the problem fixed by this commit:
1. Open the Sequences window
2. Click the Edit button to edit user-defined sequences
3. Add a new sequence
4. Save and exist the editor
5. Click the Reload button
6. Try using the new sequence. It should work.
7. Search for the new sequence. It doesn't show up.