Remove the “Disable” and “Restart” features.
Disabling WinCompose does not make much sense; the compose feature
may be disabled by selecting no compose key in the options. The user
may also choose not to start WinCompose at startup.

Restarting WinCompose only makes sense when the interface language
was changed, and there is already a button for this in the Options
window. Having a “Restart” entry in the notification icon menu will
mislead the user into thinking that whatever problem they are having
can be solved by restarting WinCompose.