Update the README's discussion of early exits
The README stated that Compose q led to an early exit.  As @MaaaxiKing has pointed out, this statement is now outdated.  The statement is no longer correct.

In truth, Compose q no longer leads to an early exit.  But Compose Q still does.

This commit updates the README to use a capital Q instead of a lowercase q.